AI Microservices/prod_requirement.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.0.0
numpy==2.1.2
scikit-learn==1.5.2
joblib==1.4.2